+++ CHANGELOG.md
@@ -4,8 +4,12 @@
 
 * Remove tokens from to selection if it is selected on from and vice versa [#3107](https://github.com/MyEtherWallet/MyEtherWallet/pull/3107)
 * Add swap token switch [#3103](https://github.com/MyEtherWallet/MyEtherWallet/pull/3103)
+* Fix ENS icons [#3110](https://github.com/MyEtherWallet/MyEtherWallet/pull/3110)
+* Fix Swap store BTC on ETH message [#3106](https://github.com/MyEtherWallet/MyEtherWallet/pull/3106)
 * Fix mew-select border issue [#3091](https://github.com/MyEtherWallet/MyEtherWallet/pull/3091)
 * Add user msg for swap [#3087](https://github.com/MyEtherWallet/MyEtherWallet/pull/3087)
+* Fix ui issues for aave [#3079](https://github.com/MyEtherWallet/MyEtherWallet/pull/3079)
+* Add top right button [#3081](https://github.com/MyEtherWallet/MyEtherWallet/pull/3081)
 * Recalculate rates when gas price is switched [#3039](https://github.com/MyEtherWallet/MyEtherWallet/pull/3039)
 * Fetch swap quotes only when all fields are in place [#3040](https://github.com/MyEtherWallet/MyEtherWallet/pull/3040)
 * swap updates [#3034](https://github.com/MyEtherWallet/MyEtherWallet/pull/3034)